requirements.txt
litellm
pyyaml
feedparser
requests
python-dotenv
